+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.1.0
+
+* Adopt [bpostlethwaite/colormap](https://github.com/bpostlethwaite/colormap). 
+  - New `markdown-header-coloring.colormapConfig` setting
+  
+* Add user defined settings for each header level
+  - New `markdown-header-coloring.userDefinedHeaderColor` setting
+
 ## 0.0.5
 
 * again: To reduce the CPU load, do not generate coloring arrays